feat: combinations of lemmatized tokens in entity_linking and optional templates in kb_answer_parser
