+++ deeppavlov/core/common/registry.json
@@ -31,9 +31,6 @@
   "elmo_file_paths_iterator": "deeppavlov.dataset_iterators.elmo_file_paths_iterator:ELMoFilePathsIterator",
   "elmo_model": "deeppavlov.models.elmo.elmo:ELMo",
   "emb_mat_assembler": "deeppavlov.models.preprocessors.assemble_embeddings_matrix:EmbeddingsMatrixAssembler",
-  "entity_linking": "deeppavlov.models.kbqa.entity_linking_freebase:EntityLinking",
-  "entity_linking_wikidata": "deeppavlov.models.kbqa.entity_linking_wikidata:EntityLinkingWikidata",
-  "entity_linking_wikidata_inv_ind": "deeppavlov.models.kbqa.entity_linking_wikidata_inv_ind:EntityLinkingWikidata",
   "faq_reader": "deeppavlov.dataset_readers.faq_reader:FaqDatasetReader",
   "fasttext": "deeppavlov.models.embedders.fasttext_embedder:FasttextEmbedder",
   "featurized_tracker": "deeppavlov.models.go_bot.tracker:FeaturizedTracker",
@@ -44,9 +41,7 @@
   "go_bot": "deeppavlov.models.go_bot.network:GoalOrientedBot",
   "hashing_tfidf_vectorizer": "deeppavlov.models.vectorizers.hashing_tfidf_vectorizer:HashingTfIdfVectorizer",
   "insurance_reader": "deeppavlov.dataset_readers.insurance_reader:InsuranceReader",
-  "kb_answer_parser": "deeppavlov.models.kbqa.kb_answer_parser:KBAnswerParser",
   "kb_answer_parser_wikidata": "deeppavlov.models.kbqa.kb_answer_parser_wikidata:KBAnswerParserWikidata",
-  "kb_answer_parser_wikidata_templates": "deeppavlov.models.kbqa.kb_answer_parser_wikidata_templates:KBAnswerParserWikidata",
   "kbqa_reader": "deeppavlov.dataset_readers.kbqa_reader:Conll2003DatasetReader",
   "kenlm_elector": "deeppavlov.models.spelling_correction.electors.kenlm_elector:KenlmElector",
   "keras_classification_model": "deeppavlov.models.classifiers.keras_classification_model:KerasClassificationModel",
@@ -59,7 +54,6 @@
   "logit_ranker": "deeppavlov.models.doc_retrieval.logit_ranker:LogitRanker",
   "lowercase_preprocessor": "deeppavlov.models.preprocessors.capitalization:LowercasePreprocessor",
   "mask": "deeppavlov.models.preprocessors.mask:Mask",
-  "max_n": "deeppavlov.models.kbqa.max_n:Max_n",
   "morpho_tagger": "deeppavlov.models.morpho_tagger.network:MorphoTagger",
   "morphotagger_dataset": "deeppavlov.dataset_iterators.morphotagger_iterator:MorphoTaggerDatasetIterator",
   "morphotagger_dataset_reader": "deeppavlov.dataset_readers.morphotagging_dataset_reader:MorphotaggerDatasetReader",
@@ -101,7 +95,6 @@
   "spelling_error_model": "deeppavlov.models.spelling_correction.brillmoore.error_model:ErrorModel",
   "spelling_levenshtein": "deeppavlov.models.spelling_correction.levenshtein.searcher_component:LevenshteinSearcherComponent",
   "split_tokenizer": "deeppavlov.models.tokenizers.split_tokenizer:SplitTokenizer",
-  "sq_reader": "deeppavlov.dataset_readers.sq_reader:OntonotesReader",
   "sqlite_database": "deeppavlov.core.data.sqlite_database:Sqlite3Database",
   "sqlite_iterator": "deeppavlov.dataset_iterators.sqlite_iterator:SQLiteDataIterator",
   "squad_ans_postprocessor": "deeppavlov.models.preprocessors.squad_preprocessor:SquadAnsPostprocessor",
